Moved basic matchers to their own file
